Alots of changes. struggled with getting jest to work with import. Got tests written for actions, moving on to more tests.
