--- babel.rc
@@ -1,4 +1,0 @@
-{
-  "presets": ["env", "react"],
-  "plugins": ["styled-components"]
-}
