+++ package.json
@@ -7,6 +7,9 @@
   "engines": {
     "node": ">=6.13.0"
   },
+  "transform": {
+    "^.+\\.jsx?$": "babel-jest"
+  },
   "scripts": {
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
@@ -36,9 +39,13 @@
     "url-loader": "^1.1.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.1",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
     "babel-plugin-styled-components": "^1.7.1",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "enzyme": "^3.6.0",
+    "enzyme-adapter-react-16": "^1.5.0",
     "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
@@ -46,6 +53,8 @@
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-react": "^7.11.1",
     "jest": "^23.6.0",
+    "jest-cli": "^23.6.0",
+    "jest-enzyme": "^6.0.4",
     "nodemon": "^1.18.4",
     "regenerator-runtime": "^0.12.1",
     "webpack": "^4.18.0",
